Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,10 +1,10 @@
 #include <stdio.h>
 int main(){
     int radius;
-    int area;
+    float area;
     scanf("%d" , &radius);
     area= 3.14*radius*radius;
-    printf("Area: %d\n" , area);
+    printf("Area: %.2f\n" , area);
     return 0;
 
 }
